reset pass and email timer

+++ app/Http/Controllers/AuthController.php
@@ -10,6 +10,7 @@ use App\Models\PasswordResetToken;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Str;
 use App\Mail\ResetPass;
+use Carbon\Carbon;
 
 
 
@@ -74,44 +75,59 @@ class AuthController extends Controller
         return view('auth.pass');
     }
     // lupa password prosses
+
     public function lupaPasswordProses(Request $request)
     {
-        // validate innput email
+        // Validate input email
         $validatedData = $request->validate([
             'email' => 'required|email',
             'nama' => 'required',
-        ]
-        ,[
+        ], [
             'email.required' => 'Email harus diisi',
             'email.email' => 'Email tidak valid',
             'nama.required' => 'Nama harus diisi',
         ]);
-        // check if user email and name is exist in user
+
+        // Check if user email and name exist in the user table
         $user = User::where('email', $validatedData['email'])->where('name', $validatedData['nama'])->first();
-        if ($user){
-            // check if email and name is exist in password reset token
-            $emailExist = PasswordResetToken::where('email', $validatedData['email'])->where('nama', $validatedData['nama'])->first();
-            if ($emailExist){
-                // delete token
+
+        if ($user) {
+            // Check if email and name exist in password reset token
+            $emailExist = PasswordResetToken::where('email', $validatedData['email'])
+            ->where('nama', $validatedData['nama'])
+            ->first();
+
+            if ($emailExist) {
+                // Check if the token was generated within the last 5 minutes
+                $tokenCreatedAt = Carbon::parse($emailExist->created_at);
+                $currentDateTime = Carbon::now();
+
+                if ($tokenCreatedAt->diffInMinutes($currentDateTime) <= 5) {
+                    return redirect()->route('lupa.password')->with('error', 'Permintaan reset password sudah dilakukan dalam 5 menit terakhir');
+                }
+
+                // Delete the previous token
                 $emailExist->delete();
             }
-            // create token
+
+            // Create a new token
             $token = new PasswordResetToken();
             $token->email = $validatedData['email'];
             $token->nama = $validatedData['nama'];
             $token->token = Str::random(40);
             $token->save();
-            // access token
-            $token = $token->token;
-            $nama = $validatedData['nama'];
-            $resetEmail = new ResetPass($token, $nama);
+
+            // Send the reset password email
+            $resetEmail = new ResetPass($token->token, $validatedData['nama']);
             Mail::to($validatedData['email'])->send($resetEmail);
-            // redirect
+
+            // Redirect
             return redirect()->route('lupa.password')->with('success', 'Email telah dikirim');
         } else {
             return redirect()->route('lupa.password')->with('error', 'Email atau nama tidak ditemukan');
         }
     }
+
     // reset password
     public function resetPassword($token, $nama)
     {
@@ -128,7 +144,7 @@ class AuthController extends Controller
     {
         // validate input password
         $validatedData = $request->validate([
-            'newpass' => 'required|min:8',
+            'newpass' => 'required|regex:/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).{8,}$/',
             'confirmpass' => 'required|same:newpass',
         ]
         ,[
